third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-reporting-for-Element-outerHTML-expected.txt

This is a testharness.js-based test.
[FAIL] No violation reported for TrustedHTML.
  assert_equals: no exception thrown expected null but got object "NoModificationAllowedError: Failed to set the 'outerHTML' property on 'Element': This element has no parent node."
Harness: the test ran to completion.
